AI_ML/src/requirements.txt
numpy
scipy
librosa
openai-whisper
sentence-transformers
scikit-learn
ffmpeg-python
google-genai
pydantic

